+++ NEWS.md
@@ -3,6 +3,7 @@
 * added `reciprocity_cor()`
 * fixed wrong str print (#5)
 * switched from Simulated Annealing to Genetic Algorithm (#4)
+* added more tests
 
 # netUtils 0.7.0