+++ NEWS.md
@@ -4,6 +4,7 @@
 * fixed wrong str print (#5)
 * switched from Simulated Annealing to Genetic Algorithm (#4)
 * added more tests
+* added `sample_lfr()` (#9)
 
 # netUtils 0.7.0